+++ src/core/data/middlewares/attach-mock/interface.ts
@@ -6,6 +6,17 @@
  * https://github.com/V4Fire/Core/blob/master/LICENSE
  */
 
+import type { Mock } from 'core/data';
+import type { MiddlewareParams } from 'core/request';
+
 export interface MockOptions {
 	patterns: RegExp[];
+}
+
+export type RequestMatchingData = Partial<Pick<MiddlewareParams['opts'], 'query' | 'body' | 'headers'>>;
+
+export interface MockBestMatch {
+	score: number;
+	mismatches: number;
+	mock: Mock | null;
 }