+++ src/config/CHANGELOG.md
@@ -9,6 +9,18 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.95.1 (2023-04-18)
+
+#### :rocket: New Feature
+
+* Added the ability to change the data storage path for the `node` runtime `core/kv-storage`
+
+## v3.92.0 (2023-02-21)
+
+#### :boom: Breaking Change
+
+* New default config for `core/net`
+
 ## v3.70.2 (2021-12-07)
 
 #### :bug: Bug Fix