+++ src/core/request/CHANGELOG.md
@@ -9,6 +9,24 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.93.1 (2023-03-14)
+
+#### :bug: Bug Fix
+
+* Fixed Node.js engine
+
+## v3.93.0 (2023-03-14)
+
+#### :rocket: New Feature
+
+* Added a new option `forceResponseType`
+
+## v3.89.1 (2023-01-19)
+
+#### :bug: Bug Fix
+
+* Fixed memory leak when caching requests with a given ttl
+
 ## v3.81.2 (2022-04-04)
 
 #### :bug: Bug Fix